wire v2: the region reference slot is EIGHT bytes (owner: "move now")
A self-relative delta cannot reach further than the width it is stored in, so
a four-byte slot bounded ONE REGION at 2 GiB — a ceiling a mesh or texture
catalogue is exactly the thing to meet. The owner ruled it on #251 today, over
the two facts already on the page: "u64 now, why fuck around" for wire v2's
widths, and "Assume we have say, 100mbs or many gigabytes of data in
Assets.bin at some point." for the scale.

ir's layout model, the Go cooker, the checker, uncook and the generator all
widen; the emitted TableRef goes uint32_t -> int64_t, which also fixes a
latent bug beside it: its comment claimed region deltas are always positive,
which §6.3 refutes for every shared node's back-reference, while its own deref
already sign-extended.

The reach refusal and §15's three widening roads are deleted as moot. §6.3
states the cost, §13.4 records the ruling with the verbatims, and §7.3's
worked cook is re-derived to the byte against real output: 200 bytes now, not
168, with Node at sizeof=16 and Scene at sizeof=32.

+++ internal/codegen/cpptable/arena.go
@@ -59,12 +59,21 @@ static const int32_t kTableMaxDepth = 128;
 //                   so a deref is one add, needs no base pointer, and a whole
 //                   region relocates by memcpy with zero fix-up
 //
-// 0 is null in both. Region deltas are always POSITIVE: a region is packed by
-// a depth-first walk, so a child always sits after the slot that names it —
-// which is what makes a packed region cycle-free by construction.
+// 0 is null in both, and a slot can never name the node that contains it, so
+// zero names nothing real in either form.
+//
+// A REGION DELTA HAS NO REQUIRED SIGN (§6.3). A region is packed depth-first,
+// so a node's FIRST reference points forward; every LATER reference to that
+// same node points BACK at the one body it already has, which is exactly what
+// makes one node one node in a region. Sharing and a back-reference are the
+// same fact, and nothing validates a reference by its sign.
+//
+// IT IS EIGHT BYTES, SIGNED, so ONE REGION REACHES EVERYTHING (§6.3, §7): a
+// four-byte slot bounded a region at 2 GiB, and the scale a cook exists for is
+// *"100mbs or many gigabytes of data in Assets.bin"*.
 struct TableRef
 {
-    uint32_t value = 0;
+    int64_t value = 0;
     bool null() const { return value == 0; }
 };
